refactor(separate): 불필요한 추상 클래스를 제거
- 단위 테스트를 위해 작성된 추상 클래스는 오버엔지니어링이라고 생각 됨
- 추상 클래스를 제거하고 Fake 객체는 Prompt를 상속 받으면 단위 테스트를 할 수 있음
