+++ src/main/java/calculator/application/CalculatorService.java
@@ -1,13 +1,12 @@
 package calculator.application;
 
 import calculator.domain.Calculator;
-import calculator.domain.prompt.CalculatorPrompt;
-import calculator.domain.prompt.Prompt;
+import calculator.domain.Prompt;
 
 public class CalculatorService {
 
     public int splitAndSum(String command) {
-        Prompt prompt = new CalculatorPrompt(command);
+        Prompt prompt = new Prompt(command);
         return new Calculator(prompt).sum();
     }
