+++ src/main/java/calculator/domain/Prompt.java
@@ -1,19 +1,24 @@
-package calculator.domain.prompt;
+package calculator.domain;
 
+import calculator.utility.CharacterUtils;
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 
-public class CalculatorPrompt extends Prompt {
+public class Prompt {
 
     private static final String CUSTOM_DELIMITER_SEPARATOR_PREFIX = "//";
     private static final String CUSTOM_DELIMITER_SEPARATOR_SUFFIX = "\\\\n";
     private static final int EXPECTED_SEPARATE_LENGTH = 2;
 
-    public CalculatorPrompt(String inputData) {
-        super(inputData);
+    private final Delimiter delimiter;
+    private final List<String> input;
+
+    public Prompt(String inputData) {
+        this.delimiter = new Delimiter();
+        this.input = parseInput(inputData);
     }
 
-    @Override
     protected List<String> parseInput(String inputData) {
         if (inputData == null || inputData.isEmpty()) {
             return new ArrayList<>();
@@ -39,7 +44,6 @@ public class CalculatorPrompt extends Prompt {
         }
     }
 
-    @Override
     protected List<String> separate(String inputData) {
         List<String> separatedData = new ArrayList<>();
         StringBuilder stringBuilder = new StringBuilder();
@@ -55,6 +59,41 @@ public class CalculatorPrompt extends Prompt {
         separatedData.add(stringBuilder.toString());
 
         return separatedData;
+    }
+
+    public List<Integer> extractInteger() {
+        return this.getInput()
+                .stream()
+                .map(this::parseNumber)
+                .toList();
+    }
+
+    private Integer parseNumber(String data) {
+        validateNumber(data);
+        return Integer.valueOf(data);
+    }
+
+    private void validateNumber(String data) {
+        validateNotNullOrEmpty(data);
+        validateOnlyDigits(data);
+    }
+
+    private void validateOnlyDigits(String data) {
+        for (char word: data.toCharArray()) {
+            if (CharacterUtils.isNotDigit(word)) {
+                throw new IllegalArgumentException();
+            }
+        }
+    }
+
+    private void validateNotNullOrEmpty(String data) {
+        if (data == null || data.isEmpty()) {
+            throw new IllegalArgumentException();
+        }
+    }
+
+    public List<String> getInput() {
+        return Collections.unmodifiableList(input);
     }
 
 }
