+++ src/test/java/calculator/domain/DelimiterTest.java
@@ -1,4 +1,4 @@
-package calculator.domain.prompt;
+package calculator.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
